- beim bohren und fräsen wird position mit config speed angefahren - Setzen von M0-2 in der steppe init

+++ CubeMx/Core/Modules/Inc/motion.h
@@ -9,12 +9,12 @@ void MoveTo(StepperMotor* motor, long target, uint32_t speed);
 
 void Internal_Line(long x_end, long y_end, uint32_t speed);
 
-void Motion_Line(long x_start, long y_start, long x_end, long y_end, uint32_t speed, double depth);
+void Motion_Line(long x_start, long y_start, long x_end, long y_end, uint32_t speed, long depth);
 
-void Motion_Triangle(long x1, long y1, long x2, long y2, long x3, long y3, uint32_t speed, double depth);
+void Motion_Triangle(long x1, long y1, long x2, long y2, long x3, long y3, uint32_t speed, long depth);
 
-void Motion_Rectangle(long x1, long y1, long x2, long y2, uint32_t speed, double depth);
+void Motion_Rectangle(long x1, long y1, long x2, long y2, uint32_t speed, long depth);
 
-void Motion_Circle(long x, long y, uint32_t r, uint32_t speed, double depth);
+void Motion_Circle(long x, long y, uint32_t r, uint32_t speed, long depth);
 
 #endif
